renaming shared files to utils.ts
